sdio: Add support for lock-less interrupt handlers *EXPERIMENTAL*
SDIO interrupt handlers are called with the host being claimed. Some
drivers may normally claim the host after acquiring a different lock.
This means that the interrupt handler may not acquire this lock, as the
locking order will be incorrect.

Add support for lock-less interrupt handlers, which will be called
without the host being claimed.

Signed-off-by: Ido Reis <idor@ti.com>
Signed-off-by: Eliad Peller <eliad@wizery.com>